wgengine/magicsock: tests call conn.bind.Close()
https://github.com/tailscale/tailscale/pull/3014 added a
rebind on STUN failure, which means there can now be a
tailscale.com/wgengine/magicsock.(*RebindingUDPConn).ReadFromNetaddr
in progress at the end of the test, waiting for a STUN
response which will never arrive.

This causes a test failure due to the resource leak. For
whatever reason, it mostly fails with Windows.

The WG scaffolding to call bind.Close() is not present in this
these tests, so ensure it gets called explicitly.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>